Remove vcr, webmock, and VCR cassettes
Currently, testing this admin app locally requires having a local instance of the API running. Therefore, it doesn't make sense to use VCR because the location IDs will be different for everyone. VCR stores the HTTP request in a cassette and reuses it instead of making a live API call. But if it doesn't recognize the HTTP request (i.e. it has different parameters than those stored in the cassette), it will raise an error.

This commit also consolidates the Warden login helpers into spec_helper and session_helpers