Make category checkboxes dynamic to reflect current state
Before, we were using static html that was generated from the Open Eligibility hierarchy and with input values set to the category IDs returned by Ohana API. It wasn't looking up the current service's categories to see which ones it already has.

This commit refactors the cats2html rake task so that it generates dynamic HAML.

Other changes:
<label> tag is now on its own, like this:
```
<ul>
  <li class="depth0">
    <input id="category-emergency">
    <label>Emergency</label>
  </li>
</ul>
```
Before, the input was inside the label, which would cause the checking/unchecking to fail in Firefox.

Added code to the init() function to check on load if any checkboxes are checked, and if so, automatically reveal them.