Add service fields to "add a new location" form
Updated the fields on the "new location" form with the descriptions from the "update location" form

Added service fields so that the service gets automatically created for the location right after the location itself is created.

The new location form automatically includes the organization's URL by picking out one of the URLs from the other locations. This should really be coming from the organization's URL attribute, which doesn't exist currently, but should. The reason for pre-populating the URL is because the list of locations that a user sees when signed in is currently based on the domain name of either the location's URL or the general email address. If the user doesn't fill out the general email field or the URL field, then they won't see this new location after it is created. This whole hack won't be necessary once we add an "admin" attribute for an organization that will be used to determine who gets to manage an organization's locations.

Increased the maxlength of certain fields and updated the specs for those fields.

Added a new rake task for creating the "new location" version of the categories partial.