Fix Heroku setup script
The DEVISE env var was misnamed, and the api endpoint was misnamed in the usage instructions.

I also removed the call to script/users to allow people to create their own users from scratch as opposed to pre-populating their production DB with dummy data.

The script now enforces the presence of each option and argument.