Update Ohanakapa to 1.1.0 but keep Faraday < 0.9
Faraday 0.9 introduced a bug with the way nil values are interpreted. The cilent expects to be able to send a nil field to specify that it should be set to an empty array. For example, to delete all emails for a Location, you should be able to send { emails: nil }, and the API knows to set the emails field to an empty array. With Faraday 0.9, it changes "nil" to an empty string which has a different meaning and will raise a Validation error.

See this PR: https://github.com/lostisland/faraday/pull/345